+++ portfolio/src/main/java/com/revature/portfolio/controllers/ProjectController.java
@@ -1,6 +1,9 @@
 package com.revature.portfolio.controllers;
 
 
+import com.revature.portfolio.PortfolioApplication;
+import com.revature.portfolio.jwt.JwtTokenUtil;
+import com.revature.portfolio.models.Developer;
 import com.revature.portfolio.models.Project;
 import com.revature.portfolio.services.ProjectService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -30,12 +33,24 @@ public class ProjectController {
 
 
     @PostMapping(value = "/Projects", consumes = "application/json", produces = "application/json")
-    public Project addProject(@RequestBody Project project) {
+    public Project addProject(@RequestBody Project project, @RequestHeader("Authorization") String header) {
+        // Get authorization header and validate
+        final String token = header.split(" ")[1].trim();
+        JwtTokenUtil tokenUtil = PortfolioApplication.app.getBean(JwtTokenUtil.class);
+        if(token == null || !tokenUtil.getUsername(token).equals(project.getDevUsername()) && !tokenUtil.getUsername(token).equals("admin"))
+            return null;
+
         return service.add(project);
     }
 
     @PutMapping(value = "/Projects/{id}", consumes = "application/json", produces = "application/json")
-    public Project updateProject(@PathVariable("id") String id, @RequestBody Project project) {
+    public Project updateProject(@PathVariable("id") String id, @RequestBody Project project,
+                                 @RequestHeader("Authorization") String header) {
+        // Get authorization header and validate
+        final String token = header.split(" ")[1].trim();
+        JwtTokenUtil tokenUtil = PortfolioApplication.app.getBean(JwtTokenUtil.class);
+        if(token == null || !tokenUtil.getUsername(token).equals(project.getDevUsername()) && !tokenUtil.getUsername(token).equals("admin"))
+            return null;
 
         project.setId(Integer.parseInt(id));
         return service.update(project);
@@ -43,7 +58,16 @@ public class ProjectController {
 
 
     @DeleteMapping("/Projects/{id}")
-    public ResponseEntity<Project> deleteProject(@PathVariable("id") String id) {
+    public ResponseEntity<Project> deleteProject(@PathVariable("id") String id,
+                                                 @RequestHeader("Authorization") String header) {
+        Project project = service.get(Integer.parseInt(id));
+
+        // Get authorization header and validate
+        final String token = header.split(" ")[1].trim();
+        JwtTokenUtil tokenUtil = PortfolioApplication.app.getBean(JwtTokenUtil.class);
+        if(token == null || !tokenUtil.getUsername(token).equals(project.getDevUsername()) && !tokenUtil.getUsername(token).equals("admin"))
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+
         boolean success = service.delete(Integer.parseInt(id));
         return new ResponseEntity<>((success) ? HttpStatus.NO_CONTENT : HttpStatus.NOT_FOUND);
     }
